Move defaultclicommands to pkg
