+++ cmd/asciichgolangpublic/cmd/root.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
-	"github.com/asciich/asciichgolangpublic/defaultclicommands"
+	"github.com/asciich/asciichgolangpublic/pkg/defaultclicommands"
 	"github.com/asciich/asciichgolangpublic/pkg/mustutils"
 )
