+++ pkg/defaultclicommands/kubernetescmd/ListKindNamesCmd.go
